Added db/user.go
